requirements.txt
aiohttp==3.8.1
beautifulsoup4==4.10.0
graia-scheduler==0.0.7
graia-ariadne==0.6.6
graia-saya==0.0.15
graiax-silkcoder==0.1.6
fastapi==0.75.0
imageio==2.16.1
jieba==0.42.1
loguru==0.6.0
matplotlib==3.5.1
moviepy==1.0.3
numpy==1.22.3
Pillow==9.0.1
pydantic==1.9.0
python_dateutil==2.8.2
PyYAML==6.0
qrcode==7.3.1
sqlalchemy==1.4.32
tencentcloud_sdk_python==3.0.596
uvicorn==0.17.5
wordcloud==1.8.1
aiosqlite==0.17.0
alembic==1.7.6
playwright==1.19.1
pytz==2021.3
psutil==5.9.0
pyzipper==0.3.5
bs4==0.0.1
torch==1.10.2
realesrgan==0.2.4.0
basicsr==1.3.5
more-itertools
youtube_dl